trading_companion/requirements.txt
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
cryptography>=42.0.0
ddgs>=0.1.0
